Hello World!
This is a squashed commit of my way to the first public version
of this IndexView plugin.

I have squashed the way because I didn't had pay much attention to
the commit messages or the logical grouped content of each commit.

To document a minimum of history and to honor the previous developers
is this not the first but the second commit.