CppParser: Group function definitions by name space
Or is the right term class name? When you have a bigger file with more than
one class can it be a little bit difficult to find the correct function in the
tree.

They got always an own root node and will not placed below some possible
present 'namespace foo {' definition, in the hope it's more handy.