Allow to set column too, not only the line number in a node
With that (bigger) change can by click on a node the cursor not only set
to the line but also to the column of the found symbol in the document.

That is in most cases not so important but with bad formatted documents
like XML or JSON is that quickly mandatory to be useful at all.

Furthermore is the addNode function removed from Parser master class, because
not needed there and some other parser required a return of bool if successful.