DocumentParser: Don't treat single char lines as empty
..but ignore them

The intent not to clutter the tree with useless entries didn't work (now?)
with this approach. But when the line is ignored it does.

Because DocumentParser is a master class without info page I hike up the
version number of the PlaintextParser.

The enum changes may not required but now looks nicer.