IndexView: Remove range check in update item to skip update
This check is sadly too aggressive and skip also an update where one is
desired. When we have a nested situation, like a class in header file,
and the current item is the top class item, then is never the item
updated when moving the cursor inside the class :-/

Considering that this plugin do much more work on each edit, is this little
optimization approach somehow pointless. So, note at me: Nice try.