The (not so) final flicker-avoidance II
This addition blocks the update of the tree in a more sane way and
avoid the flicker on doc switch.

I think the current anti flicker stuff works so far nice but it is somehow
ugly implemented. Guess we should keep a tree per doc. But no idea how
expensive this is.