Cache for each doc the tree and parser
This makes doc switching delay free and avoid as best as possible any unpleasant
flickering when editing.

I did some poor test to estimate how much memory this way may be consumed and it
seemed that it take only ~20KB for a pretty BIG file with 35k lines. So I guess
even with 100 open files should we not run into trouble with nowadays systems.

- ConfigPage: Show info for each loaded parser class
- Protect against a mass flood of doc change signals
- Only parse when needed, never forget an update
- Fix/Apply Parser::docType()
- Remove Parser(s) in a sane way
- Remove view option actions from IndexView, ask Parser instead
- Remove many quirks to avoid flicker
- Rename (S&R) docChanged to viewChanged
- Add doc as parameter to slot docModeChanged