Add detection of git conflict blocks
This patch aboard the parsing when a git conflict block was found, clears the
tree, search for more possible conflict blocks and only show theses in the index.

Furthermore is the context menu disabled, the header set to a hint and the view
mode temporary changed to plain regardless of user setting.

Possible ToDo/Change:
  - Disable the update of the tree until user saved the file