Add a lookup button right from the filter field
This is like "filter on steroids" from our own "auto-filter" functionality when
you select some text or "symbol lookup" or "goto declaration" known from other
tools. How useful it will be, I don't know!

As said, select some text/word and hit the button. Then will be performed a search
in all open files, more precisely in our cached trees. If something match will that
shown as kind of tree. If nothing was found, no more happens as to indicate that in
a red colored filter field.

When you click some item in the tree will the current view switch to that doc and
place the cursor as usual.

Hint: The current document is skiped in that search, just to keep the results free
      from useless offers.