ProgramParser: Fix false detection of FIXME/BEGIN in heredoc
This fix is more a demolition than a repair, but I can't remember nor
find a hint why it was done so complicated. The true fix should have been
done in stripLine() by checking if we are in a heredoc...

But at the moment I see no need for that overkill. Thanks to git, we
should be able to restore the old stuff if necessary.