Add preprocessor directive to generate a parser status report
To enable this stuff you need to edit the #define above Parser::parser()

Then will be reports placed in KatePlugin-IndexView/tests/reports
when you view a file from KatePlugin-IndexView/tests

The magic take place when we track these reports by git. This way we can
see if we lost some node when we fumble with the regex or such.

This is my workaround for my inability (or laziness) to implement real tests.